pyproject.toml
[project]
name = "pyalexatodo"
version = "0.1.0"
description = "Library to access and edit Alexa To-do and Shoppings list through the inofficial Alexa API."
readme = "README.md"
authors = [
    { name = "lonlazer", email = "decoy-dwelled-onto@duck.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aioamazondevices>=13.4.2",
    "pydantic>=2.10.6",
]

[project.urls]
Homepage = "https://github.com/lonlazer/pyalexatodo"
Issues = "https://github.com/lonlazer/pyalexatodo/issue"

[project.scripts]
pyalexatodo = "pyalexatodo.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "commitizen>=4.13.10",
    "pdoc>=16.0.0",
    "pytest>=8.0.0",
    "pytest-aiohttp>=1.1.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.15.11",
    "ty>=0.0.31",
]

[project.optional-dependencies]
cli = [
    "keyring>=25.6.0",
    "typer>=0.15.2",
]